Move helper class for stdout/stderr to top
